Fix load by alternate key with Automatic primary key
Storables ignore @Automatic properties when checking if the primary key
is initialized. This makes sense for insert, because the Repository is
responsible for populating the @Automatic property. However, this doesn't
work for load--when the @Automatic property is also the primary key--
because it won't attempt to load by alternate key when it must.